Problem: objects produced by reducers are not documented
This means they have implicit, poorly understood schemas.

Solution: designate a repository to collect "most known",
public schemas to be published on sit-it.org

Why a separate repository? Aside from being able to
use GitHub Pages functionality to publish them (this will
benefit will eventually go away because we'd want to
move hosting of sit-it.org to something that can do
SSL [Let's Encrypt] natively), the real reason for this
separate is to maintain the independence of "contracts"
(schemas) and their implementations.